+++ src/api/consumer/index.js
@@ -17,7 +17,7 @@ import getComponentLogs from './lib/get-component-logs';
 import { add as remoteAdd, list as remoteList, remove as remoteRm } from './lib/remote';
 import config from './lib/global-config';
 import getDriver from './lib/get-driver';
-import { watchAll } from './lib/watch';
+import watchAll from './lib/watch';
 import add from './lib/add';
 import untrack from './lib/untrack';
 import { unTagAction, unTagAllAction } from './lib/untag';