+++ src/api/consumer/lib/untag.js
@@ -1,6 +1,7 @@
 // @flow
 import { loadConsumer, Consumer } from '../../../consumer';
 import { BitId } from '../../../bit-id';
+import { removeLocalVersion, removeLocalVersionsForAllComponents } from '../../../scope/component-ops/untag-component';
 
 export async function unTagAction(id: string, version?: string, force: boolean) {
   const consumer: Consumer = await loadConsumer();
@@ -9,11 +10,11 @@ export async function unTagAction(id: string, version?: string, force: boolean) 
   // a user might run the command `bit untag id@version` instead of `bit untag id version`
   if (bitId.hasVersion() && !version) version = bitId.version;
 
-  const result = await consumer.scope.removeLocalVersion(bitId, version, force);
+  const result = await removeLocalVersion(consumer.scope, bitId, version, force);
   return [result];
 }
 
 export async function unTagAllAction(version?: string, force: boolean) {
   const consumer: Consumer = await loadConsumer();
-  return consumer.scope.removeLocalVersionsForAllComponents(version, force);
+  return removeLocalVersionsForAllComponents(consumer.scope, version, force);
 }