+++ src/consumer/consumer.js
@@ -68,7 +68,8 @@ type ComponentStatus = {
   newlyCreated: boolean,
   deleted: boolean,
   staged: boolean,
-  notExist: boolean
+  notExist: boolean,
+  nested: boolean // when a component is nested, it doesn't matter whether it was modified
 };
 
 export default class Consumer {
@@ -626,8 +627,23 @@ export default class Consumer {
         status.newlyCreated = true;
         return status;
       }
+      if (componentFromFileSystem.componentMap.origin === COMPONENT_ORIGINS.NESTED) {
+        status.nested = true;
+        return status;
+      }
       status.staged = componentFromModel.isLocallyChanged();
       const versionFromFs = componentFromFileSystem.id.version;
+      if (!status.staged && !componentFromFileSystem.id.hasVersion()) {
+        throw new Error(
+          `component ${id} has an invalid state.
+           1) it has a model instance so it's not new.
+           2) it's not in staged state.
+           3) it doesn't have a version in the bitmap file.
+           Maybe the component was interrupted during the export and as a result the bitmap file wasn't updated with the new version
+           `
+        );
+      }
+
       const latestVersionFromModel = componentFromModel.latest();
       // Consider the following two scenarios:
       // 1) a user tagged v1, exported, then tagged v2.
@@ -637,6 +653,7 @@ export default class Consumer {
       //    @see reduce-path.e2e 'importing v1 of a component when a component has v2' to reproduce this case.
       const version = status.staged ? latestVersionFromModel : versionFromFs;
       const versionRef = componentFromModel.versions[version];
+      if (!versionRef) throw new Error(`version ${version} was not found in ${id}`);
       const versionFromModel = await this.scope.getObject(versionRef.hash);
       status.modified = await this.isComponentModified(versionFromModel, componentFromFileSystem);
       return status;
@@ -756,7 +773,7 @@ export default class Consumer {
     let resolvedScopePath = path.join(projectPath, BIT_HIDDEN_DIR);
     // let addedGitHooks;
     let existingGitHooks;
-    if (!noGit && fs.existsSync(gitDirPath)) {
+    if (!noGit && fs.existsSync(gitDirPath) && !fs.existsSync(resolvedScopePath)) {
       resolvedScopePath = path.join(gitDirPath, BIT_GIT_DIR);
       // const gitHooksManager = GitHooksManager.init(gitDirPath);
       // const writeHooksResult = gitHooksManager.writeAllHooks();
@@ -1063,11 +1080,12 @@ export default class Consumer {
     return this.installPackages(componentDirs, verbose);
   }
 
-  async eject(componentsIds) {
-    const componentIdsWithoutScope = componentsIds.map(id => id.toStringWithoutScope());
-    await this.remove(componentIdsWithoutScope, true, false, true);
+  async eject(componentsIds: BitId[]) {
     await packageJson.addComponentsWithVersionToRoot(this, componentsIds);
+    const componentIdsWithoutScope = componentsIds.map(id => id.toStringWithoutScope());
+    await packageJson.removeComponentsFromNodeModules(this, componentsIds);
     await this.installPackages([], true, true);
+    await this.remove(componentIdsWithoutScope, true, false, true);
     return componentsIds;
   }
 }