+++ src/utils/fs-read-file.js
@@ -1,7 +1,7 @@
 /** @flow */
 import fs from 'fs';
 
-export default function readFile(path: string) {
+export default function readFile(path: string): Promise<Buffer> {
   return new Promise((resolve, reject) => {
     fs.readFile(path, (err, res) => {
       if (err) return reject(err);