+++ CHANGELOG.md
@@ -9,10 +9,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - support passing arguments/flags to the package-manager by specifying them after '--' (e.g. bit import -- --no-optional)
 
+## [0.12.6-dev.5] - 2018-02-20
+
+- bug fix - untagging component with dependents and local tags after export wasn't prevented as it should
+- bug fix - untagging component with dependencies was showing an error saying the component is depends on itself
+
+## [0.12.6-dev.4] - 2018-02-19
+
+- save bit dev-dependencies components inside devDependencies section of package.json
+- fix bit cc to clear module's cache
+- fix resolving .gitignore files
+- when adding existing files to component in bitmap don't create duplications
+
 ## [0.12.6-dev.3] - 2018-02-15
 
 - support `bit install` of specific ids
 - support peerDependencies (in the model for now)
+- [#729](https://github.com/teambit/bit/issues/729) fix bit cc to clear module cache
+- [#769](https://github.com/teambit/bit/issues/769) - prevent duplicate ids in bitmap when adding existing files
 
 ## [0.12.6-dev.2] - 2018-02-13
 
@@ -22,6 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - introduced a new command `bit untag` for reverting un-exported tags.
 - init local scope inside .git
 - support vue files
+- [#736](https://github.com/teambit/bit/issues/736) - .gitignore is blocking everything
 
 ## [0.12.5] - 2018-02-06